requirements.txt
pybluez
paho-mqtt

